Having to use auto detection of platforms to specify JS
