+++ package.json
@@ -13,7 +13,6 @@
     "email": "support@pspdfkit.com"
   },
   "license": "SEE LICENSE IN LICENSE",
-  "main": "js/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/PSPDFKit/react-native.git"
